feat: add base ncurses bindings and stubs
Beginning with this commit, we'll be using this style of external
function calls. By using a class to encompass the particular calls,
we can easily derive and/or mock those classes in testing.

Here, we bring in:

    - src/ncurses.hpp
        - Should be included by all application code interfacing with
          ncurses. This allows application code to switch based on the
          presence of the -DTESTING flag during compilation.
    - src/ext/ncurses.{hpp,cpp}
        - Real ncurses wrapper
    - src/stubs/ncurses.{hpp,cpp}
        - Stubbed ncurses wrapper, can be used as a base for gmock.

We also change some meson details in terms of testing flags and
building a stub dependency out of our new stub source code,
which all tests depend on.

Signed-off-by: Kevin Morris <kevr@0cost.org>