feat: add cfg::config, a configuration class
cfg::config::ref([cfg::config&]) can be used to set and/or get
a singleton reference to a cfg::config, initialized statically
to a default-constructed cfg::config.

One should first initialize the config somewhere according
to their intent:

    auto &conf = cfg::config::ref();
    conf.option("some-var", po::value<std::string>(), "help for some-var");
    conf.parse_args(argc, argv);

Standard usage of the new config:

    auto &conf = cfg::config::ref();
    if (conf.exists("some-var")) {
        // Explicitly typecasting via conf.get<T>
        std::cout << conf.get<std::string>("some-var") << std::endl;

        // Or, with operator[] to simplify the call and force a string
        std::cout << conf["some-var"] << std::endl;
    }

At this point, only the command-line portion of the config is used.
Upcoming, we will need to implement handling of config files using
this config class with `cfg::config::parse_config(...)`

Signed-off-by: Kevin Morris <kevr@0cost.org>