feat: implement --config and fallback to {~/.config,/etc}/tasker/config
The main process now reads a config as well as command line arguments.

The new `--config arg` program option lets a user specify a config
elsewhere to load. When --config is not provided, the program searches
the following directories for a `config` file, in order:

1. `$HOME/.config/tasker/config`
2. `/etc/tasker/config`

This kind of fallback loading is still in place for most of our
tests. I'm working on a way to get all of tests to use a
fake config somewhere (like tmpdir) by default instead.

Signed-off-by: Kevin Morris <kevr@0cost.org>