feat: add locale support
This project intends to support various utf-8 characters that can
be used for display. Our binary is already linked to ncursesw,
which offers wide character support. In order for ncursesw to
truly support wide characters, a proper locale with those wide
characters needs to be set.

In this case, we set the application's locale to the system
locale configured for the user running it.

Signed-off-by: Kevin Morris <kevr@0cost.org>